requirements.txt
pywin32
Pillow
numpy
onnxruntime
imagehash